+++ src/complex/rect/rect.src.js
@@ -1,10 +1,5 @@
-import {atan2, sin, hypot, abs as stdAbs} from "../../standard/standard.lib.js";
+import {atan2, hypot} from "../../standard/standard.lib.js";
 
-function conj(z, target = new Float64Array(2)) {
-	target[0] = z[0];
-	target[1] = -z[1];
-	return target;
-}
 
 function real(z) {
 	return z[0];
@@ -20,6 +15,18 @@ function arg(z) {
 
 function abs(z) {
 	return hypot(z[0], z[1]);
+}
+
+function conj(z, target = new Float64Array(2)) {
+	target[0] = z[0];
+	target[1] = -z[1];
+	return target;
+}
+
+function negate(z, target = new Float64Array(2)) {
+	target[0] = -z[0];
+	target[1] = -z[1];
+	return target;
 }
 
 function add(z1, z2, target = new Float64Array(2)) {
@@ -40,7 +47,7 @@ function mult(z1, z2, target = new Float64Array(2)) {
 	const re2 = z2[0];
 	const im2 = z2[1];
 	target[0] = re1 * re2 - im1 * im2;
-	target[1] = re1 * im2 - re2 * im1;
+	target[1] = re1 * im2 + re2 * im1;
 	return target;
 }
 
@@ -55,18 +62,22 @@ function div(z1, z2, target = new Float64Array(2)) {
 	const im1 = z1[1];
 	const re2 = z2[0];
 	const im2 = z2[1];
-	const scale = 1 / (re2 * re2 + im2 * im2);
-	target[0] = (re1 * re2 + im1 * im2) * scale;
-	target[1] = (- re1 * im2 - re2 * im1) * scale;
+	const mag2Squared = re2 * re2 + im2 * im2;
+	if (mag2Squared === 0) {return undefined;}
+	const scaleFactor = 1 / mag2Squared;
+	target[0] = (re1 * re2 + im1 * im2) * scaleFactor;
+	target[1] = (- re1 * im2 - re2 * im1) * scaleFactor;
 	return target;
 }
 
 function inverse(z, target = new Float64Array(2)) {
 	const re = z[0];
 	const im = z[1];
-	const scale = 1 / (re * re + im * im);
-	target[0] = (re * re + im * im) * scale;
-	target[1] = (- re * im - re * im) * scale;
+	const magSquared = re * re + im * im;
+	if (magSquared === 0) {return undefined;}
+	const scaleFactor = 1 / magSquared;
+	target[0] = re * scaleFactor;
+	target[1] = -im * scaleFactor;
 	return target;
 }
 
@@ -92,6 +103,7 @@ Object.freeze(scale);
 Object.freeze(div);
 Object.freeze(inverse);
 Object.freeze(toPolar);
+Object.freeze(negate);
 
 // Export
-export {conj, real, imag, arg, abs, add, sub, mult, scale, div, inverse, toPolar}
+export {conj, real, imag, arg, abs, add, sub, mult, scale, div, inverse, toPolar, negate}